Commented out code that prevents successful run
I have removed part of the testing code, since it prevented a successful run through of the code. This means testing if the displaying magnetisation function works properly is possible just by running it once. I also changed up how the strain and stress functions are related. Now you just input a generic strain into the stress function, which means that including the magnetostrain will be as simple as taking a difference within the stress function.